Check for error state/clear in SOM reading
The std::getline can trigger an error state when attempting to locate the end of a header by simply reading the entire file searching for the string.

Clearing the error is required before the seekg and tellg calls make sense and work as expected. Without doing so there will be readSOM errors thrown