Only scan for end of header if header implied
Scanning an entire file for the `# END OF HEADER` string can be very time consuming if large inputing files (i.e. image binary) are used. For small files there is nearly no delay, but for files that are 10s+GB it can take minutes to begin training/mapping.

This will only scan for the end of a header if there is a header implied by the first character being a #, otherwise no scan is performed.