#Fix 23: --inter-store keep not working
It was a bug in the ProgressBar. The valid function was not working,
because next_progress_print was updated within the increment operator.
This works for printing the progress bar but not for the later valid()
function, which is used for writing intermediate results.

Fixed by change from comparison with next number to modulus.

Added some unit tests and output stream.